refactor(api)!: arity is a count, so make it usize
arity and param_count are parameter counts and cannot be negative; the
descriptor validation said as much by rejecting arity < 0. Unsigned says
it in the type instead, matching argc.

That leaves the range unchecked at the other end, which mattered more:
bytecode encodes a call's argument count in a single byte, so an arity
above 255 wrapped silently when the call was emitted. OAK_MAX_ARITY makes
the ceiling explicit and oak_bind_fn/oak_bind_fn_global reject anything
past it -- including a count mistakenly written as a negative int, which
now lands there rather than being registered as an enormous one.

The compiler's own registries stay signed: the implicit self is a -1/+1
adjustment throughout them, and that arithmetic is safer signed. The
descriptor's count converts once, at the registration boundary.

Breaking: arity and param_count on both bind descriptors, and on
oak_obj_fn_t / oak_obj_native_fn_t, are usize.

+++ include/oak_value.h
@@ -504,7 +504,7 @@ struct oak_obj_fn
 {
   oak_obj_t obj;
   usize code_offset;
-  int arity;
+  usize arity;
   u16 module_id;
   const char* name;
   oak_attr_hook_entry_t* attr_hooks;
@@ -556,7 +556,7 @@ struct oak_obj_native_fn
 {
   oak_obj_t obj;
   oak_native_fn_t fn;
-  int arity;
+  usize arity;
   const char* name;
   /* Forwarded to the callback through oak_native_call_t::user_data. */
   void* user_data;
@@ -753,12 +753,12 @@ OAK_API oak_obj_string_t* oak_string_concat(oak_allocator_t* a,
 
 OAK_API oak_obj_fn_t* oak_fn_new(oak_allocator_t* a,
                                  usize code_offset,
-                                 int arity,
+                                 usize arity,
                                  u16 module_id);
 
 OAK_API oak_obj_native_fn_t* oak_native_fn_new(oak_allocator_t* a,
                                                oak_native_fn_t fn,
-                                               int arity,
+                                               usize arity,
                                                const char* name,
                                                void* user_data);
